refactor: consolidate SVG cropping logic into crop_svg_sprite.py; delete new_crop.py
