+++ src/animations_sprites/crop_svg_sprite.py
@@ -1,463 +1,329 @@
-import os
-import re
-import shutil
-import tempfile
-import subprocess
-import xml.etree.ElementTree as ET
-from copy import deepcopy
+from __future__ import annotations
+
+import io
 from pathlib import Path
+from typing import List, Tuple, Iterable
+import xml.etree.ElementTree as ET
+
+import numpy as np
+from PIL import Image
+import cairosvg
 
 from src.utils.config import SPRITES_SHEETS_DIR, SPRITES_POSES_DIR
 
-INPUT_FILE = Path(SPRITES_SHEETS_DIR) / "adventurer_vector.svg"
-OUTPUT_DIR = Path(SPRITES_POSES_DIR)
+INPUT_FILE = SPRITES_SHEETS_DIR / "player_vector.svg"
+OUTPUT_DIR = SPRITES_POSES_DIR
 PREFIX = "pose"
 
+# Parameters
+RASTER_SCALE = 1.0
+INTERSECT_PAD = 1.0
+EXPORT_MARGIN = 8.0
+MIN_GROUPS_IN_POSE = 4
+MIN_POSE_W, MIN_POSE_H = 20.0, 20.0
 
-SVG_NS = "http://www.w3.org/2000/svg"
-XLINK_NS = "http://www.w3.org/1999/xlink"
-ET.register_namespace("", SVG_NS)
-ET.register_namespace("xlink", XLINK_NS)
-
-# Clustering sensitivity:
-# Minimum absolute intersection area (px^2) to link two groups
-MIN_INTERSECTION_AREA = 100
-# Minimum intersection share vs. the smaller bbox to link two groups
-MIN_INTERSECTION_PERCENT = 0.10
-
-
-def tag(local: str) -> str:
-    return f"{{{SVG_NS}}}{local}"
+# Namespaces
+NS_SVG = "http://www.w3.org/2000/svg"
+NS_XLINK = "http://www.w3.org/1999/xlink"
+ET.register_namespace("", NS_SVG)
+ET.register_namespace("xlink", NS_XLINK)
 
 
-def tagn(el: ET.Element) -> str:
-    return el.tag.split("}")[-1]
+def _localname(tag: str) -> str:
+    """Return tag without namespace prefix."""
+    return tag.split("}", 1)[1] if tag.startswith("{") else tag
 
 
-def _find_inkscape_binary() -> str:
-    """
-    Locate Inkscape binary (env -> PATH -> common install locations).
-    """
-    env_bin = os.environ.get("INKSCAPE_BIN")
-    if env_bin and Path(env_bin).exists():
-        return env_bin
-
-    for name in ("inkscape", "inkscape.com"):
-        path = shutil.which(name)
-        if path:
-            return path
-
-    candidates = [
-        "/Applications/Inkscape.app/Contents/MacOS/inkscape",  # macOS
-        r"C:\Program Files\Inkscape\bin\inkscape.com",  # Windows
-        r"C:\Program Files\Inkscape\bin\inkscape.exe",
-        r"C:\Program Files (x86)\Inkscape\bin\inkscape.com",
-        r"C:\Program Files (x86)\Inkscape\bin\inkscape.exe",
-        "/usr/bin/inkscape",  # Linux
-        "/usr/local/bin/inkscape",
-    ]
-    for path in candidates:
-        if Path(path).exists():
-            return path
-
-    raise RuntimeError("Inkscape not found. Install it or set INKSCAPE_BIN.")
+def _parse_svg(svg_path: Path) -> ET.Element:
+    """Parse SVG and return the root element."""
+    return ET.parse(svg_path).getroot()
 
 
-def mat_identity():
-    """Return identity 2x3 affine matrix [a b c d e f]."""
-    return [1.0, 0.0, 0.0, 1.0, 0.0, 0.0]
+def _get_canvas_size(root: ET.Element) -> Tuple[float, float]:
+    """Get canvas width/height from viewBox or width/height attributes."""
+    vb = root.get("viewBox")
+    if vb:
+        _, _, w, h = map(float, vb.split())
+        return w, h
+
+    def _num(s: str | None) -> float:
+        return float(str(s).replace("px", "")) if s else 1000.0
+
+    return _num(root.get("width")), _num(root.get("height"))
 
 
-def mat_mul(A, B):
-    """Matrix multiply: A∘B (apply B first, then A)."""
-    a1, b1, c1, d1, e1, f1 = A
-    a2, b2, c2, d2, e2, f2 = B
+def _collect_top_groups(root: ET.Element) -> List[Tuple[int, ET.Element]]:
+    """Collect top-level <g> elements only."""
     return [
-        a1 * a2 + c1 * b2,
-        b1 * a2 + d1 * b2,
-        a1 * c2 + c1 * d2,
-        b1 * c2 + d1 * d2,
-        a1 * e2 + c1 * f2 + e1,
-        b1 * e2 + d1 * f2 + f1,
+        (idx, child) for idx, child in enumerate(root) if _localname(child.tag) == "g"
     ]
 
 
-def mat_translate(tx, ty):
-    """Translation matrix."""
-    return [1.0, 0.0, 0.0, 1.0, float(tx), float(ty)]
+def _build_svg_wrapper(
+    children: Iterable[ET.Element],
+    view_box: Tuple[float, float, float, float],
+    copy_defs_from: ET.Element | None,
+) -> str:
+    """Wrap given children into a standalone SVG string."""
+    x, y, w, h = view_box
+    attrs = {
+        "xmlns": NS_SVG,
+        "xmlns:xlink": NS_XLINK,
+        "viewBox": f"{x} {y} {w} {h}",
+        "width": str(w),
+        "height": str(h),
+    }
+    attr_str = " ".join(f'{k}="{v}"' for k, v in attrs.items())
+    parts = [f"<svg {attr_str}>"]
+
+    if copy_defs_from is not None:
+        defs = copy_defs_from.find(f".//{{{NS_SVG}}}defs")
+        if defs is not None:
+            parts.append(ET.tostring(defs, encoding="unicode"))
+
+    parts.append("<g>")
+    parts.extend(ET.tostring(el, encoding="unicode") for el in children)
+    parts.append("</g></svg>")
+    return "".join(parts)
 
 
-def mat_from_transform(attr: str):
-    """
-    Parse `transform` attribute into a 2x3 matrix.
-    Supports `translate(...)` and `matrix(...)` (add scale/rotate if needed).
-    """
-    M = mat_identity()
-    if not attr:
-        return M
+def _build_single_group_svg(
+    group: ET.Element,
+    base_w: float,
+    base_h: float,
+    copy_defs_from: ET.Element | None,
+) -> str:
+    """Wrap one group into full-canvas SVG."""
+    attrs = {
+        "xmlns": NS_SVG,
+        "xmlns:xlink": NS_XLINK,
+        "viewBox": f"0 0 {base_w} {base_h}",
+        "width": str(int(base_w)),
+        "height": str(int(base_h)),
+    }
+    attr_str = " ".join(f'{k}="{v}"' for k, v in attrs.items())
+    parts = [f"<svg {attr_str}>"]
 
-    for match in re.finditer(r"(matrix|translate)\s*\(([^)]*)\)", attr):
-        kind = match.group(1)
-        args = [float(x) for x in re.split(r"[, \t]+", match.group(2).strip()) if x]
-        if kind == "translate":
-            tx = args[0] if args else 0.0
-            ty = args[1] if len(args) > 1 else 0.0
-            M = mat_mul(M, mat_translate(tx, ty))
-        elif kind == "matrix" and len(args) == 6:
-            M = mat_mul(M, args)
-    return M
+    if copy_defs_from is not None:
+        defs = copy_defs_from.find(f".//{{{NS_SVG}}}defs")
+        if defs is not None:
+            parts.append(ET.tostring(defs, encoding="unicode"))
 
-
-def apply_mat(M, x, y):
-    """Apply 2x3 matrix to a point (x, y)."""
-    a, b, c, d, e, f = M
-    return (a * x + c * y + e, b * x + d * y + f)
-
-
-def bbox_union(A, B):
-    """Union of two bboxes (x1, y1, x2, y2)."""
-    if A is None:
-        return B
-    if B is None:
-        return A
-    return (min(A[0], B[0]), min(A[1], B[1]), max(A[2], B[2]), max(A[3], B[3]))
+    parts.append(ET.tostring(group, encoding="unicode"))
+    parts.append("</svg>")
+    return "".join(parts)
 
 
-def bbox_area(bbox):
-    """Area of bbox."""
-    if bbox is None:
-        return 0.0
-    x1, y1, x2, y2 = bbox
-    return max(0.0, x2 - x1) * max(0.0, y2 - y1)
+def _raster_bbox_of_group(
+    group: ET.Element,
+    base_w: float,
+    base_h: float,
+    defs_root: ET.Element | None,
+    scale: float = RASTER_SCALE,
+    alpha_threshold: int = 1,
+) -> Tuple[float, float, float, float] | None:
+    """Rasterize a single <g> and compute visible bbox."""
+    svg_str = _build_single_group_svg(group, base_w, base_h, defs_root)
+    png_bytes = cairosvg.svg2png(
+        bytestring=svg_str.encode("utf-8"),
+        output_width=int(base_w * scale),
+        output_height=int(base_h * scale),
+    )
 
+    if png_bytes is None:
+        raise RuntimeError("cairosvg.svg2png returned None")
 
-def bbox_intersection(A, B):
-    """Intersection of two bboxes, or None if disjoint."""
-    if A is None or B is None:
+    img = Image.open(io.BytesIO(png_bytes)).convert("RGBA")
+    alpha = np.array(img)[..., 3]
+    ys, xs = np.where(alpha > alpha_threshold)
+    if xs.size == 0:
         return None
-    ax1, ay1, ax2, ay2 = A
-    bx1, by1, bx2, by2 = B
-    x1 = max(ax1, bx1)
-    y1 = max(ay1, by1)
-    x2 = min(ax2, bx2)
-    y2 = min(ay2, by2)
-    if x1 >= x2 or y1 >= y2:
-        return None
-    return (x1, y1, x2, y2)
+
+    xmin_px, xmax_px = xs.min(), xs.max()
+    ymin_px, ymax_px = ys.min(), ys.max()
+    return (
+        xmin_px / scale,
+        ymin_px / scale,
+        (xmax_px + 1) / scale,
+        (ymax_px + 1) / scale,
+    )
 
 
-def should_cluster_together(bbox1, bbox2):
-    """
-    Decide if two bboxes should be linked in the same cluster.
-    Requires both a minimum absolute area and a minimum relative share
-    w.r.t. the smaller bbox.
-    """
-    inter = bbox_intersection(bbox1, bbox2)
-    if inter is None:
-        return False
-
-    inter_area = bbox_area(inter)
-    if inter_area < MIN_INTERSECTION_AREA:
-        return False
-
-    area1 = bbox_area(bbox1)
-    area2 = bbox_area(bbox2)
-    smaller = min(area1, area2)
-    if smaller > 0:
-        share = inter_area / smaller
-        return share >= MIN_INTERSECTION_PERCENT
-    return False
+def _rects_intersect(
+    a: Tuple[float, float, float, float],
+    b: Tuple[float, float, float, float],
+    pad: float = 0.0,
+) -> bool:
+    """Check if two rectangles intersect with optional padding."""
+    ax0, ay0, ax1, ay1 = a
+    bx0, by0, bx1, by1 = b
+    return not (
+        ax1 <= bx0 - pad or bx1 <= ax0 - pad or ay1 <= by0 - pad or by1 <= ay0 - pad
+    )
 
 
-def path_points_from_d(d: str):
-    """
-    Extract control points from path 'd'. Handles M/L/H/V/C/Q/S/T (abs/rel).
-    For curves we include control points and end points to over-approximate bbox.
-    """
-    if not d:
-        return []
-
-    tokens = re.findall(r"([MmLlHhVvCcQqSsTtZz])|([-+]?\d*\.?\d+(?:[eE][-+]?\d+)?)", d)
-    points = []
-    current_pos = (0.0, 0.0)
-
-    command_groups = []
-    current_command = None
-    current_numbers = []
-
-    for cmd_flag, number in tokens:
-        if cmd_flag:
-            if current_command and current_numbers:
-                command_groups.append((current_command, current_numbers))
-            current_command = cmd_flag
-            current_numbers = []
-        elif number:
-            current_numbers.append(float(number))
-
-    if current_command and current_numbers:
-        command_groups.append((current_command, current_numbers))
-
-    def add_point(x, y, relative=False):
-        nonlocal current_pos
-        if relative:
-            current_pos = (current_pos[0] + x, current_pos[1] + y)
-        else:
-            current_pos = (x, y)
-        points.append(current_pos)
-
-    for command, numbers in command_groups:
-        if not command:
-            continue
-
-        is_rel = command.islower()
-        cmd = command.upper()
-        vals = numbers
-        i = 0
-
-        while i < len(vals):
-            if cmd in ("M", "L", "T") and i + 1 < len(vals):
-                add_point(vals[i], vals[i + 1], relative=is_rel)
-                i += 2
-            elif cmd == "H" and i < len(vals):
-                add_point(vals[i], current_pos[1], relative=is_rel)
-                i += 1
-            elif cmd == "V" and i < len(vals):
-                add_point(current_pos[0], vals[i], relative=is_rel)
-                i += 1
-            elif cmd == "C" and i + 5 < len(vals):
-                if is_rel:
-                    points.extend(
-                        [
-                            (current_pos[0] + vals[i], current_pos[1] + vals[i + 1]),
-                            (
-                                current_pos[0] + vals[i + 2],
-                                current_pos[1] + vals[i + 3],
-                            ),
-                        ]
-                    )
-                    add_point(vals[i + 4], vals[i + 5], relative=True)
-                else:
-                    points.extend([(vals[i], vals[i + 1]), (vals[i + 2], vals[i + 3])])
-                    add_point(vals[i + 4], vals[i + 5], relative=False)
-                i += 6
-            elif cmd in ("Q", "S") and i + 3 < len(vals):
-                if is_rel:
-                    points.append(
-                        (current_pos[0] + vals[i], current_pos[1] + vals[i + 1])
-                    )
-                    add_point(vals[i + 2], vals[i + 3], relative=True)
-                else:
-                    points.append((vals[i], vals[i + 1]))
-                    add_point(vals[i + 2], vals[i + 3], relative=False)
-                i += 4
-            else:
-                break
-
-    return points
+def _union_bbox(
+    rects: List[Tuple[float, float, float, float]],
+) -> Tuple[float, float, float, float]:
+    """Compute union bbox for a list of rectangles."""
+    xs = [coord for rect in rects for coord in (rect[0], rect[2])]
+    ys = [coord for rect in rects for coord in (rect[1], rect[3])]
+    return min(xs), min(ys), max(xs), max(ys)
 
 
-def bbox_of_path_d(d: str, transform_matrix):
-    """Compute a path bbox after applying the transform matrix."""
-    points = path_points_from_d(d)
-    if not points:
-        return None
-    transformed = [apply_mat(transform_matrix, x, y) for x, y in points]
-    xs, ys = zip(*transformed)
-    return (min(xs), min(ys), max(xs), max(ys))
-
-
-def build_id_map(defs: ET.Element) -> dict[str, ET.Element]:
-    """Build id -> element map for <defs> symbols/groups."""
-    id_map = {}
-    if defs is None:
-        return id_map
-    for element in defs.iter():
-        element_id = element.attrib.get("id")
-        if element_id:
-            id_map[element_id] = element
-    return id_map
-
-
-def calculate_bbox_recursive(
-    element: ET.Element, parent_matrix, id_map
-) -> tuple | None:
-    """
-    Recursively compute a bbox for an element considering transforms.
-    Supports <g>, <use>, and <path>.
-    """
-    bbox = None
-    name = tagn(element)
-
-    if name == "g":
-        local = mat_mul(parent_matrix, mat_from_transform(element.get("transform", "")))
-        for child in element:
-            child_bbox = calculate_bbox_recursive(child, local, id_map)
-            bbox = bbox_union(bbox, child_bbox)
-
-    elif name == "use":
-        href = element.attrib.get(f"{{{XLINK_NS}}}href")
-        if href and href.startswith("#"):
-            target = id_map.get(href[1:])
-            if target is not None:
-                local = mat_mul(
-                    parent_matrix, mat_from_transform(element.get("transform", ""))
-                )
-                bbox = calculate_bbox_recursive(target, local, id_map)
-
-    elif name == "path":
-        local = mat_mul(parent_matrix, mat_from_transform(element.get("transform", "")))
-        bbox = bbox_of_path_d(element.get("d", ""), local)
-
-    return bbox
-
-
-def build_clusters(items):
-    """
-    Cluster elements by intersections of their bboxes.
-    items: list[(ET.Element, bbox)]
-    returns: list[list[ET.Element]]
-    """
-    n = len(items)
-    if n == 0:
-        return []
-
-    parent = list(range(n))
-
-    def find(x):
-        if parent[x] != x:
-            parent[x] = find(parent[x])
-        return parent[x]
-
-    def union(x, y):
-        rx, ry = find(x), find(y)
-        if rx != ry:
-            parent[rx] = ry
-
+def _cluster_groups_into_poses(
+    groups: List[Tuple[int, ET.Element]],
+    bboxes: List[Tuple[float, float, float, float]],
+    pad: float = INTERSECT_PAD,
+) -> List[List[int]]:
+    """Build intersection graph and return connected components."""
+    n = len(groups)
+    adj = [[] for _ in range(n)]
     for i in range(n):
         for j in range(i + 1, n):
-            if should_cluster_together(items[i][1], items[j][1]):
-                union(i, j)
+            if _rects_intersect(bboxes[i], bboxes[j], pad):
+                adj[i].append(j)
+                adj[j].append(i)
 
-    buckets = {}
+    seen = [False] * n
+    components = []
     for i in range(n):
-        r = find(i)
-        buckets.setdefault(r, []).append(items[i][0])
+        if seen[i]:
+            continue
+        stack = [i]
+        seen[i] = True
+        comp = []
+        while stack:
+            u = stack.pop()
+            comp.append(u)
+            for v in adj[u]:
+                if not seen[v]:
+                    seen[v] = True
+                    stack.append(v)
+        components.append(comp)
+    return components
 
-    return list(buckets.values())
+
+def _filter_and_sort_components(
+    components: List[List[int]],
+    bboxes: List[Tuple[float, float, float, float]],
+) -> List[List[int]]:
+    """Drop tiny components and sort poses in row-major order."""
+    valid = []
+    for comp in components:
+        if len(comp) < MIN_GROUPS_IN_POSE:
+            continue
+        x0, y0, x1, y1 = _union_bbox([bboxes[i] for i in comp])
+        if (x1 - x0) < MIN_POSE_W or (y1 - y0) < MIN_POSE_H:
+            continue
+        valid.append(comp)
+
+    def _row_major_key(c: List[int]) -> tuple:
+        x0, y0, x1, y1 = _union_bbox([bboxes[i] for i in c])
+        return (round(y0 // 100), x0)
+
+    return sorted(valid, key=_row_major_key)
 
 
-def export_with_inkscape(
-    inkscape_binary: str, source_svg: Path, output_svg: Path
-) -> None:
-    """
-    Export SVG via Inkscape cropped to drawing bounds.
-    """
-    command = [
-        inkscape_binary,
-        "--export-type=svg",
-        "--export-plain-svg",
-        "--export-area-drawing",
-        f"--export-filename={output_svg}",
-        str(source_svg),
-    ]
-    result = subprocess.run(command, capture_output=True, text=True)
-    if result.returncode != 0:
-        raise RuntimeError(
-            f"Inkscape export failed:\nSTDOUT: {result.stdout}\nSTDERR: {result.stderr}"
-        )
+def _export_pose_svgs(
+    root: ET.Element,
+    groups: List[Tuple[int, ET.Element]],
+    components: List[List[int]],
+    bboxes: List[Tuple[float, float, float, float]],
+    out_dir: Path,
+    prefix: str = "pose",
+) -> List[Path]:
+    """Export each component as a standalone SVG with tight viewBox."""
+    out_dir.mkdir(parents=True, exist_ok=True)
+    base_w, base_h = _get_canvas_size(root)
+
+    exported = []
+    for k, comp in enumerate(components, start=1):
+        x0, y0, x1, y1 = _union_bbox([bboxes[i] for i in comp])
+        x0 = max(0.0, x0 - EXPORT_MARGIN)
+        y0 = max(0.0, y0 - EXPORT_MARGIN)
+        x1 = min(base_w, x1 + EXPORT_MARGIN)
+        y1 = min(base_h, y1 + EXPORT_MARGIN)
+        w, h = max(1.0, x1 - x0), max(1.0, y1 - y0)
+
+        comp_sorted = sorted(comp, key=lambda i: groups[i][0])
+        children = (groups[i][1] for i in comp_sorted)
+
+        svg_str = _build_svg_wrapper(children, (x0, y0, w, h), root)
+        out_path = out_dir / f"{prefix}_{k:02d}.svg"
+        out_path.write_text(svg_str, encoding="utf-8")
+        exported.append(out_path)
+
+    return exported
 
 
 def get_cropped_poses(
     input_file: str | Path = INPUT_FILE,
     output_dir: str | Path = OUTPUT_DIR,
     prefix: str = PREFIX,
-):
+) -> List[Path]:
     """
-    Crop poses from a single SVG sprite sheet.
+    Smart sprite cropping entry point.
 
     Args:
-        input_file (str | Path, optional): The input SVG file to crop. Defaults to INPUT_FILE.
-        output_dir (str | Path, optional): The directory to save cropped poses. Defaults to OUTPUT_DIR.
+        input_file (str | Path, optional): The input SVG file to process. Defaults to INPUT_FILE.
+        output_dir (str | Path, optional): The directory to save output SVG files. Defaults to OUTPUT_DIR.
         prefix (str, optional): The prefix for output file names. Defaults to PREFIX.
 
     Raises:
         FileNotFoundError: If the input file is not found.
-        RuntimeError: If no valid poses are found.
+        FileNotFoundError: If the input directory is not found.
+        RuntimeError: If no drawable top-level <g> elements are detected.
+        RuntimeError: If no valid poses are detected after clustering.
 
-    End-to-end pipeline:
-    - parse input SVG file
-    - build id map from <defs>
-    - collect root <g> groups and compute bboxes
-    - cluster groups by intersection rule
-    - build minimal SVG per cluster and export via Inkscape
+    Returns:
+        List[Path]: A list of output SVG file paths.
     """
     input_file = Path(input_file)
     output_dir = Path(output_dir)
 
     if not input_file.exists():
-        new_input = Path(SPRITES_SHEETS_DIR) / input_file
-        if new_input.exists():
-            input_file = new_input
-        else:
-            raise FileNotFoundError(f"Input file not found: {input_file} and {new_input}")
+        raise FileNotFoundError(f"Input file not found: {input_file}")
 
-    output_dir.mkdir(parents=True, exist_ok=True)
+    # check if input file is directory
+    if input_file.is_dir():
+        print(f"Input is a directory. Processing first SVG file found in it.")
+        svg_files = list(input_file.glob("*.svg"))
+        if not svg_files:
+            raise FileNotFoundError(f"No SVG files found in directory: {input_file}")
 
-    tree = ET.parse(input_file)
-    root = tree.getroot()
+        input_file = svg_files[0]
 
-    defs = next((child for child in root if child.tag == tag("defs")), None)
-    id_map = build_id_map(defs) if defs is not None else {}
 
-    root_groups = [el for el in root if tagn(el) == "g"]
-    items = []
-    for g in root_groups:
-        bbox = calculate_bbox_recursive(g, mat_identity(), id_map)
-        if bbox is not None:
-            items.append((g, bbox))
+    root = _parse_svg(input_file)
+    base_w, base_h = _get_canvas_size(root)
+    top_groups = _collect_top_groups(root)
 
-    if not items:
-        raise RuntimeError("No root groups with geometry were found.")
+    bboxes = []
+    valid_groups = []
+    for idx, g in top_groups:
+        bb = _raster_bbox_of_group(g, base_w, base_h, root, scale=RASTER_SCALE)
+        if bb is not None:
+            bboxes.append(bb)
+            valid_groups.append((idx, g))
 
-    print(f"Found {len(items)} groups with geometry")
+    if not valid_groups:
+        raise RuntimeError("No drawable top-level <g> elements detected.")
 
-    clusters = build_clusters(items)
-    print(f"Built {len(clusters)} clusters")
+    comps = _cluster_groups_into_poses(valid_groups, bboxes, pad=INTERSECT_PAD)
+    comps = _filter_and_sort_components(comps, bboxes)
 
-    inkscape_binary = _find_inkscape_binary()
+    if not comps:
+        raise RuntimeError("No valid poses detected after clustering.")
 
-    for idx, cluster_groups in enumerate(clusters, start=1):
-        new_root = ET.Element(tag("svg"))
-        for attr in ("width", "height", "viewBox"):
-            val = root.get(attr)
-            if val:
-                new_root.set(attr, val)
-
-        if defs is not None:
-            new_root.append(deepcopy(defs))
-
-        wrap = ET.SubElement(new_root, tag("g"), {"id": f"cluster_{idx:02d}"})
-        for g in cluster_groups:
-            wrap.append(deepcopy(g))
-
-        with tempfile.TemporaryDirectory() as td:
-            tmp_svg = Path(td) / f"cluster_{idx:02d}.svg"
-            ET.ElementTree(new_root).write(
-                tmp_svg, encoding="utf-8", xml_declaration=True
-            )
-
-            out_path = output_dir / f"{prefix}_{idx:02d}.svg"
-            export_with_inkscape(inkscape_binary, tmp_svg, out_path)
-            print(f"✓ {out_path.name}")
-
-    print(f"\nDone! Created {len(clusters)} files in: {output_dir}")
+    return _export_pose_svgs(root, valid_groups, comps, bboxes, output_dir, prefix)
 
 
 def get_all_cropped_poses(
     input_dir: str | Path = SPRITES_SHEETS_DIR,
     output_dir: str | Path = OUTPUT_DIR,
-    prefix: str = PREFIX,
-) -> list[Path]:
+    prefix: str = "pose",
+) -> List[Path]:
     """
     Get all cropped poses from SVG files in the input directory.
 
@@ -493,3 +359,7 @@ def get_all_cropped_poses(
         output_dirs.append(curr_output_dir)
         print("-" * 40)
     return output_dirs
+
+
+if __name__ == "__main__":
+    get_cropped_poses()
